requirements.txt
anyio>=4.0
bcrypt>=4.2.0
certifi>=2024.0
fastapi>=0.135.0
httpx>=0.28.0
jinja2>=3.1.0
openai>=2.20.0
pydantic>=2.12.0
python-dotenv>=1.0.0
python-jose>=3.5.0
python-multipart>=0.0.20
starlette>=0.52.0
supabase>=2.15.0
typing-extensions>=4.15.0
